automated.offline.server: Non-working EJBConfiguration (direct injection of EntityManager)
Moved automated.offline.server/src/main/java/.../automated/offline/server/*.java to
  misc/example-jpa/src/main/.../automated/offline/server/

Replaced BatchJobHome interfaces and implementations with direct injection of an EntityManager.

Removed or consolidated a bunch of throw(s) clauses as a part of replacing *Home interfaces
with EntityManager injections.

Changed the type of BatchJob.jobId and BatchJob.transactionId from Long to long.

+++ 2.7.x/choicemaker-cm/choicemaker-j2ee/com.choicemaker.cm.urm.ejb/src/main/java/com/choicemaker/cm/urm/ejb/BatchQueryListener.java
@@ -52,7 +52,7 @@ public class BatchQueryListener extends WorkflowControlListener{
 								CmRuntimeException,SQLException,CreateException,ArgumentException,ModelException {
 
 		BatchJob batchJob = Single.getInst().findBatchJobById(batchJobId);
-		long urmJobId = batchJob.getTransactionId().longValue();
+		long urmJobId = batchJob.getTransactionId();
 		return urmJobId;
 	}
